App is now complete.
You can now add employee leave when you need to and look at it when you want to.

The dates skip weekends and only take into account week days.

When you create an event it gets added to google calendar as an event.

You can now select the employee that you want to take leave for.