Add role base authorization and authentication.
When a user or admin logs in it shows the data that is specific to them.

Admins will have the ability to approve or not approve leave.

users can only view their leave.

nav menu only shows what is specific to ones role.

Add an approve, approvedBy and dateApproved to their respective models.